code/src/backend/requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.9.0
atlassian-python-api==3.41.21
attrs==25.3.0
beautifulsoup4==4.13.3
behave==1.2.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
dotenv==0.9.9
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langgraph-sdk==0.1.58
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
orjson==3.10.16
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
